tweaked tests
