+++ test/Sentry.Unity.Tests/IntegrationTests.cs
@@ -37,7 +37,7 @@ namespace Sentry.Unity.Tests
             // We don't have it in tests, but in scenes.
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
         }
 
@@ -52,7 +52,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -72,7 +72,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -90,7 +90,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
 
@@ -110,7 +110,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
 
@@ -128,7 +128,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -148,7 +148,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -164,7 +164,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Not.Contain(expectedAttribute));
         }
@@ -186,8 +186,8 @@ namespace Sentry.Unity.Tests
             var testBehaviour = new GameObject("TestHolder").AddComponent<TestMonoBehaviour>();
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.ThrowException), _eventMessage);
 
-            Assert.AreEqual(string.Empty, firstHttpClientHandler.GetEvent(_eventReceiveTimeout));
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            Assert.AreEqual(string.Empty, firstHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout));
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
         }
 
@@ -202,7 +202,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute)); // sanity check
             Assert.That(triggeredEvent, Does.Contain(expectedMechanism));
         }
@@ -220,7 +220,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogError), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -238,7 +238,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogErrorInTask), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -255,7 +255,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogException), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
@@ -272,7 +272,7 @@ namespace Sentry.Unity.Tests
 
             testBehaviour.gameObject.SendMessage(nameof(testBehaviour.DebugLogExceptionInTask), _eventMessage);
 
-            var triggeredEvent = _testHttpClientHandler.GetEvent(_eventReceiveTimeout);
+            var triggeredEvent = _testHttpClientHandler.GetEvent(TestEventType.SentryEvent, _eventReceiveTimeout);
             Assert.That(triggeredEvent, Does.Contain(_identifyingEventValueAttribute));
             Assert.That(triggeredEvent, Does.Contain(expectedAttribute));
         }
